Books rating system added

+++ src/Core/LibraryApp.Application/Feauters/Reviews/Commands/Create/CreateReviewCommand.cs
@@ -5,7 +5,7 @@ namespace LibraryApp.Application.Feauters.Reviews.Commands.Create
     public record CreateReviewCommand(
         Guid UserId,
         Guid BookId,
-        int Rating,
+        double Rating,
         string? Title,
         string? Text) : IRequest<Guid>;
 }
